# *: reduce call when comparing key