Signed-off-by: Ryan Leung <rleungx@gmail.com>

<!--
Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/tikv/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed
-->

### What problem does this PR solve?

<!--

Please create an issue first to describe the problem.
There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#linking-issues.

-->
Issue Number: Ref #5606.

### What is changed and how does it work?
```
go test -benchmem -count 5 -run=^$ -bench ^BenchmarkLoadRegionsByVolume$ github.com/tikv/pd/server/storage

name                                       old time/op    new time/op    delta
LoadRegionsByVolume/input_size_10000-40      0.05ns ±13%    0.04ns ±25%     ~     (p=0.056 n=5+5)
LoadRegionsByVolume/input_size_100000-40     0.60ns ± 9%    0.47ns ± 6%  -21.15%  (p=0.008 n=5+5)
LoadRegionsByVolume/input_size_1000000-40     6.84s ± 3%     5.35s ± 6%  -21.84%  (p=0.008 n=5+5)

name                                       old alloc/op   new alloc/op   delta
LoadRegionsByVolume/input_size_10000-40       0.00B          0.00B          ~     (all equal)
LoadRegionsByVolume/input_size_100000-40      0.00B          0.00B          ~     (all equal)
LoadRegionsByVolume/input_size_1000000-40     934MB ± 0%     933MB ± 0%   -0.01%  (p=0.032 n=5+5)

name                                       old allocs/op  new allocs/op  delta
LoadRegionsByVolume/input_size_10000-40        0.00           0.00          ~     (all equal)
LoadRegionsByVolume/input_size_100000-40       0.00           0.00          ~     (all equal)
LoadRegionsByVolume/input_size_1000000-40     15.2M ± 0%     15.2M ± 0%     ~     (p=0.421 n=5+5)
```
```
go test -benchmem -count 5 -run=^$ -bench ^BenchmarkLoadRegionsByRandomMerge$ github.com/tikv/pd/server/storage

name                                        old time/op    new time/op    delta
LoadRegionsByRandomMerge/merge_ratio_0-40      6.47s ± 6%     5.33s ± 5%  -17.62%  (p=0.008 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_20-40     7.99s ± 3%     7.01s ± 4%  -12.32%  (p=0.008 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_40-40     9.69s ± 5%     8.64s ± 4%  -10.86%  (p=0.008 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_60-40     11.1s ± 2%      9.9s ± 5%  -10.40%  (p=0.008 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_80-40     12.4s ± 3%     11.2s ± 5%   -9.74%  (p=0.008 n=5+5)

name                                        old alloc/op   new alloc/op   delta
LoadRegionsByRandomMerge/merge_ratio_0-40      934MB ± 0%     933MB ± 0%   -0.01%  (p=0.032 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_20-40    1.01GB ± 0%    1.01GB ± 0%     ~     (p=0.056 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_40-40    1.10GB ± 0%    1.10GB ± 0%   +0.06%  (p=0.032 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_60-40    1.17GB ± 0%    1.17GB ± 0%   +0.08%  (p=0.008 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_80-40    1.25GB ± 0%    1.25GB ± 0%   +0.09%  (p=0.016 n=5+5)

name                                        old allocs/op  new allocs/op  delta
LoadRegionsByRandomMerge/merge_ratio_0-40      15.2M ± 0%     15.2M ± 0%     ~     (p=0.151 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_20-40     18.0M ± 0%     18.0M ± 0%     ~     (p=0.421 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_40-40     20.7M ± 0%     20.7M ± 0%     ~     (p=0.151 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_60-40     23.3M ± 0%     23.3M ± 0%   +0.00%  (p=0.008 n=5+5)
LoadRegionsByRandomMerge/merge_ratio_80-40     25.8M ± 0%     25.8M ± 0%     ~     (p=0.151 n=5+5)
```

<!--

You could use the "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#format-of-the-commit-message.

-->

```commit-message
```

### Check List

<!-- Remove the items that are not applicable. -->

Tests

<!-- At least one of these tests must be included. -->

- Unit test

### Release note

<!-- A bugfix or a new feature needs a release note. If there is no need release note, just uncomment the below line. -->

```release-note
None
```